pyproject.toml
[tool.poetry]
name = "home-work-27"
version = "0.1.0"
description = ""
authors = ["Stas Kholshchevnikov <xcd85@mail.ru>"]
readme = "README.md"
packages = [{include = "home_work_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
